unified code for browser and node envs
- unified code for browser and nodejs environments
- ouput cjs and esm
- add top cjs and esm files for easy switch between cjs and esm
