+++ api-extractor.json
@@ -1,6 +1,6 @@
 {
   "$schema": "https://developer.microsoft.com/json-schemas/api-extractor/v7/api-extractor.schema.json",
-  "mainEntryPointFilePath": "<projectFolder>/dist/cjs/node/index.d.ts",
+  "mainEntryPointFilePath": "<projectFolder>/dist/cjs/index.d.ts",
   "bundledPackages": [],
   "compiler": {},
   "apiReport": {
