+++ cjs.js
@@ -1,0 +1,3 @@
+'use strict';
+
+module.exports = require('./dist/cjs/index');
