+++ docs/variables/bytesfrommeta.md
@@ -7,5 +7,5 @@ Convert meta or array of metas or a schema to bytes and compress them for on-cha
 <b>Signature:</b>
 
 ```typescript
-bytesFromMeta: (meta: object | object[] | string, schema: object | string, path?: string) => string
+bytesFromMeta: (meta: object | object[] | string, schema: object | string) => string
 ```
