+++ docs/variables/concat.md
@@ -7,5 +7,5 @@ ethers concat
 <b>Signature:</b>
 
 ```typescript
-concat: typeof ethers.utils.concat
+concat: typeof utils.concat
 ```
