+++ docs/variables/hexzeropad.md
@@ -7,5 +7,5 @@ ethers hexZeroPad
 <b>Signature:</b>
 
 ```typescript
-hexZeroPad: typeof ethers.utils.hexZeroPad
+hexZeroPad: typeof utils.hexZeroPad
 ```
