+++ docs/variables/ishexstring.md
@@ -7,5 +7,5 @@ ethers isHexString
 <b>Signature:</b>
 
 ```typescript
-isHexString: typeof ethers.utils.isHexString
+isHexString: typeof utils.isHexString
 ```
