+++ docs/variables/metafrombytes.md
@@ -7,5 +7,5 @@ Decompress and convert bytes to meta
 <b>Signature:</b>
 
 ```typescript
-metaFromBytes: (bytes: BytesLike, schema: object | string, path?: string) => any
+metaFromBytes: (bytes: BytesLike, schema: object | string) => any
 ```
