+++ package.json
@@ -1,7 +1,7 @@
 {
     "name": "@rainprotocol/rainlang",
     "description": "Rainlang standalone server, that encapsulates Rain Language Serivces (in LSP spec) and compiler for NodeJs and Browser environments",
-    "version": "1.1.1",
+    "version": "1.2.0",
     "author": "Rain Protocol",
     "license": "CAL-1.0",
     "repository": {
@@ -13,17 +13,20 @@
         "url": "https://github.com/rainprotocol/rainlang/issues"
     },
     "homepage": "http://rainprotocol.github.io/rainlang/",
-    "main": "dist/cjs/node/index.js",
-    "module": "dist/cjs/node/index.js",
-    "browser": "dist/esm/browser/index.js",
+    "main": "./dist/cjs/index.js",
+    "module": "./dist/esm/index.js",
+    "browser": "./dist/esm/index.js",
     "files": [
         "dist",
         "src",
-        "docs"
+        "docs",
+        "cjs.js",
+        "esm.js",
+        "cjs.dts",
+        "esm.d.ts",
+        "README.md",
+        "package.json"
     ],
-    "engines": {
-        "node": ">=10"
-    },
     "scripts": {
         "prepare": "npm run build",
         "build": "npm run build-cjs && npm run build-esm",
@@ -57,7 +60,7 @@
         "ajv": "^8.12.0",
         "algebra.js": "git+https://github.com/rouzwelt/algebra.js",
         "buffer": "^6.0.3",
-        "cbor": "^8.1.0",
+        "cbor-web": "https://gitpkg.now.sh/rouzwelt/node-cbor/packages/cbor-web?2023-03-22-cbor-web-dts",
         "ethers": "^5.7.2",
         "graphql": "^16.6.0",
         "graphql-request": "^5.2.0",
