--- src/browser/index.ts
@@ -1,12 +1,0 @@
-export * from './compiler/rainCompiler';
-export * from './compiler/rainDecompiler';
-export * from '../shared/compiler/expressionConfigTypes';
-export * from './parser/rainParser';
-export * from '../shared/parser/opMetaTypes';
-export * from '../shared/parser/rainParserTypes';
-export * from '../shared/rainLanguageTypes';
-export * from "./services/rainHover";
-export * from "./services/rainDiagnostics";
-export * from "./services/rainCompletion";
-export * from './rainLanguageService';
-export * from './utils';
