+++ src/compiler/expressionConfigTypes.ts
@@ -1,4 +1,4 @@
-import { BigNumberish, BytesLike } from 'ethers';
+import { BigNumberish, BytesLike } from '../utils';
 
 /**
  * @public
