+++ src/index.ts
@@ -1,0 +1,19 @@
+export * from './compiler/rainCompiler';
+export * from './compiler/rainDecompiler';
+export * from './compiler/expressionConfigTypes';
+export * from './parser/rainParser';
+export * from './parser/onChainOpMeta';
+export * from './parser/opMetaTypes';
+export * from './parser/rainParserTypes';
+export * from './rainLanguageTypes';
+export * from "./services/rainHover";
+export * from "./services/rainDiagnostics";
+export * from "./services/rainCompletion";
+export * from './rainLanguageService';
+export * from './utils';
+
+import schema from "./schema/op.meta.schema.json";
+/**
+ * @public op meta schema used for validation
+ */
+export const OpMetaSchema = schema;
